Revert "Refactor LibraryInitializer so it's thread safe. Fixes random sporadical concurrency crashes. (#15762)"
This reverts commit bfd3bb8972b7e4a9cd8487c4ab6e6583202f3259.
